refactor(tests/e2e): move `getContinueWaitTime` to common helpers
