+++ tests/e2e/helpers/testWallet.ts
@@ -6,12 +6,12 @@ import {
   type KeyInfo,
 } from '../fixtures/helpers';
 import { fillPopup, type Popup, type ConnectDetails } from '../pages/popup';
-import { getWalletInformation } from '@/shared/helpers';
-import { waitForWelcomePage } from './common';
+import { getContinueWaitTime, waitForWelcomePage } from './common';
 
 export const KEYS_PAGE_URL = `https://wallet.interledger-test.dev/settings/developer-keys`;
 export const LOGIN_PAGE_URL = `https://wallet.interledger-test.dev/auth/login?callbackUrl=%2Fsettings%2Fdeveloper-keys`;
 export const API_URL_ORIGIN = `https://api.wallet.interledger-test.dev`;
+export const DEFAULT_CONTINUE_WAIT_MS = 1000;
 
 export async function connectWallet(
   context: BrowserContext,
@@ -26,7 +26,11 @@ export async function connectWallet(
   const connectButton = await fillPopup(popup, i18n, params);
   await connectButton.click();
 
-  const continueWaitMs = await getContinueWaitTime(context, params);
+  const continueWaitMs = await getContinueWaitTime(
+    context,
+    params,
+    DEFAULT_CONTINUE_WAIT_MS,
+  );
 
   const page = await context.waitForEvent('page', (page) =>
     page.url().includes('/grant-interactions'),
@@ -40,35 +44,6 @@ export async function completeGrant(page: Page, continueWaitMs: number) {
   await waitForGrantConsentPage(page);
   await acceptGrant(page, continueWaitMs);
   await waitForWelcomePage(page);
-}
-
-export async function getContinueWaitTime(
-  context: BrowserContext,
-  params: Pick<ConnectDetails, 'walletAddressUrl'>,
-) {
-  const continueWaitMs = await (async () => {
-    const defaultWaitMs = 1001;
-    if (process.env.PW_EXPERIMENTAL_SERVICE_WORKER_NETWORK_EVENTS !== '1') {
-      return Promise.resolve(defaultWaitMs);
-    }
-    const walletInfo = await getWalletInformation(params.walletAddressUrl);
-    return await new Promise<number>((resolve) => {
-      const authServer = new URL(walletInfo.authServer).href;
-      context.on('requestfinished', async function intercept(req) {
-        if (!req.serviceWorker()) return;
-        if (new URL(req.url()).href !== authServer) return;
-
-        const res = await req.response();
-        const json = await res?.json();
-        context.off('requestfinished', intercept);
-        if (typeof json?.continue?.wait !== 'number') {
-          return resolve(defaultWaitMs);
-        }
-        return resolve(json.continue.wait * 1000);
-      });
-    });
-  })();
-  return continueWaitMs;
 }
 
 export async function waitForGrantConsentPage(page: Page) {
